@wip
Add basic admin UI

Add a basic interface for store owners to create and edit premade carts.

STATUS: simple CRUD operations in place and tested.

NEXT STEPS: add UI for adding variants to a premade cart. I like how the
taxons can be added to a product so that's a good way to start.