Enable pre-made carts
Store owners sometimes need to create URLs that will add an item to a
customer's cart and send them to checkout. This helps them create
specific types of marketing campaigns. We enable this by creating a new
model: the PremadeCart. This object has_many Variants, and has a random
alphanum token. When A visitor requests the URL (made with that token)
of the premade cart, the associated variants are added to their cart and
they are redirected to checkout.

Future improvements will include better management of pre-made cart
items, and the ability to include a promotion code.

STATUS: it's terrible but it works. By forcing it to build a new
premade_cart_variant object in the view I got it to render. Now to
remove the terrible parts and make it work.